+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Update research viewer role permissions. [#546](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/546)
+- Migrate and hide design plot fields. [#549](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/549)
 
 ## [2.16.2](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/milestone/34) 2023-09-06